Rewrite tests
